requirements.txt
absl-py==1.4.0
altgraph==0.17.3
arrow==1.2.3
binaryornot==0.4.4
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colour==0.1.5
contourpy==1.0.7
cookiecutter==2.1.1
cycler==0.11.0
fonttools==4.39.0
idna==3.4
Jinja2==3.1.2
jinja2-time==0.2.0
kiwisolver==1.4.4
line-profiler==4.0.3
llvmlite==0.39.1
MarkupSafe==2.1.2
matplotlib==3.7.1
Mesa==1.2.0
networkx==3.0
numba==0.56.4
numpy==1.23.5
opencv-python==4.7.0.72
packaging==23.0
pandas==1.5.3
path-planning @ file:///home/zartris/code/cpp/Path-Planning-Lib
path-planning-lib @ file:///home/zartris/code/cpp/Path-Planning-Lib
Pillow==9.4.0
protobuf==4.22.1
pybind11==2.10.4
pyinstaller==5.9.0
pyinstaller-hooks-contrib==2023.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
scipy==1.10.1
six==1.16.0
text-unidecode==1.3
tornado==6.2
tqdm==4.65.0
urllib3==1.26.15
